Deprecate the `tuftool refresh` command
`tuftool update` replaces `tuftool refresh` and serves the purpose
of updating a repository's metadata and has the ability to add
targets.
